chore: StoryTopHeader 공통 사용을 위해 Ui 모듈로 이동
